backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12

# Database
supabase==2.9.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.36

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI
openai==1.51.2
anthropic==0.36.2
google-generativeai==0.8.3

# HTTP
httpx==0.27.2
aiohttp==3.10.10

# Queue / Scheduler
celery==5.4.0
redis==5.1.1
kombu==5.3.7

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Utils
python-dotenv==1.0.1
tenacity==9.0.0
cachetools==5.5.0

# Social Media APIs
tweepy==4.14.0
google-api-python-client==2.147.0
google-auth-oauthlib==1.2.1
linkedin-api==2.1.5

# WhatsApp
twilio==9.3.7

# Admin / Billing / Export
stripe==11.1.1
pandas==2.2.3
openpyxl==3.1.5
gunicorn==23.0.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
